requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asyncio==3.4.3
asyncpg==0.30.0
attrs==25.2.0
bcrypt==4.3.0
blinker==1.9.0
cached-property==2.0.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
Flask==3.1.0
frozenlist==1.5.0
geojson==2.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
keyboard==0.13.5
magic-filter==1.0.12
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
orjson==3.10.16
packaging==25.0
passlib==1.7.4
pluggy==1.5.0
propcache==0.3.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
pyowm==3.3.0
PySocks==1.7.1
pytest==8.3.5
pytest-asyncio==1.0.0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
python-telegram-bot==22.0
python-weather==2.0.7
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.1
rsa==4.9.1
setuptools==76.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlalchemy-history==2.1.4
SQLAlchemy-Utils==0.41.2
sqlmodel==0.0.24
starlette==0.46.2
termcolor==2.5.0
typer==0.15.2
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.1
watchfiles==1.0.5
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.18.3
